Convert ConnectionInfo to interface
- Convert `ConnectionInfo` to interface
- Add `DefaultConnectionInfo` class in `callback` package
