+++ src/main/java/io/r2dbc/proxy/core/ConnectionInfo.java
@@ -19,111 +19,69 @@ package io.r2dbc.proxy.core;
 
 import io.r2dbc.spi.Connection;
 
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
-
 /**
- * {@link Connection} related information.
+ * Hold {@link Connection} related information.
  *
  * @author Tadaya Tsuyukubo
  */
-public class ConnectionInfo {
+public interface ConnectionInfo {
 
-    private Connection originalConnection;
+    /**
+     * Retrieve original {@link Connection}.
+     *
+     * @return connection
+     */
+    Connection getOriginalConnection();
 
-    private String connectionId;
-
-    private AtomicBoolean isClosed = new AtomicBoolean();
-
-    private AtomicInteger transactionCount = new AtomicInteger();
-
-    private AtomicInteger commitCount = new AtomicInteger();
-
-    private AtomicInteger rollbackCount = new AtomicInteger();
-
-    // TODO: may keep transaction isolation level
-
-    public Connection getOriginalConnection() {
-        return this.originalConnection;
-    }
-
-    public void setOriginalConnection(Connection originalConnection) {
-        this.originalConnection = originalConnection;
-    }
-
-    public String getConnectionId() {
-        return this.connectionId;
-    }
-
-    public void setConnectionId(String connectionId) {
-        this.connectionId = connectionId;
-    }
+    /**
+     * Get ID for the connection.
+     *
+     * @return connection ID
+     * @see io.r2dbc.proxy.callback.ConnectionIdManager
+     */
+    String getConnectionId();
 
     /**
      * Increment transaction count.
      */
-    public void incrementTransactionCount() {
-        this.transactionCount.incrementAndGet();
-    }
+    void incrementTransactionCount();
 
     /**
      * Increment commit count.
      */
-    public void incrementCommitCount() {
-        this.commitCount.incrementAndGet();
-    }
+    void incrementCommitCount();
 
     /**
      * Increment rollback count.
      */
-    public void incrementRollbackCount() {
-        this.rollbackCount.incrementAndGet();
-    }
-
+    void incrementRollbackCount();
 
     /**
      * Returns how many times {@link Connection#beginTransaction()} method is called.
      *
      * @return num of beginTransaction() method being called
      */
-    public int getTransactionCount() {
-        return this.transactionCount.get();
-    }
+    int getTransactionCount();
 
     /**
      * Returns how many times {@link Connection#commitTransaction()} method is called.
      *
      * @return num of commitTransaction method being called
      */
-    public int getCommitCount() {
-        return this.commitCount.get();
-    }
+    int getCommitCount();
 
     /**
      * Returns how many times {@link Connection#rollbackTransaction()} method is called.
      *
      * @return num of rollback methods being called
      */
-    public int getRollbackCount() {
-        return this.rollbackCount.get();
-    }
+    int getRollbackCount();
 
     /**
      * Returns whether connection is closed or not.
      *
      * @return {@code true} if connection is closed
      */
-    public boolean isClosed() {
-        return this.isClosed.get();
-    }
-
-    /**
-     * Indicate connection is closed.
-     *
-     * @param closed closed
-     */
-    public void setClosed(boolean closed) {
-        this.isClosed.set(closed);
-    }
+    boolean isClosed();
 
 }
